requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
bidict==0.22.1
black==23.3.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.95.0
fastapi-jwt-auth==0.5.0
fastapi-socketio==0.0.10
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.6
PyJWT==1.7.1
python-dotenv==1.0.0
python-engineio==4.4.0
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.8.0
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
